More refactoring; BaseHoodieMergedLogRecordScanner expose getRecords and reuse builders
