+++ hudi-cli/src/main/java/org/apache/hudi/cli/commands/HoodieLogFileCommand.java
@@ -225,7 +225,7 @@ public class HoodieLogFileCommand {
     if (shouldMerge) {
       System.out.println("===========================> MERGING RECORDS <===================");
       HoodieMergedLogRecordScanner scanner =
-          HoodieMergedLogRecordScanner.newBuilder()
+          (HoodieMergedLogRecordScanner) HoodieMergedLogRecordScanner.newBuilder()
               .withStorage(storage)
               .withBasePath(client.getBasePath())
               .withLogFilePaths(logFilePaths)
