+++ hudi-integ-test/src/main/java/org/apache/hudi/integ/testsuite/reader/DFSHoodieDatasetInputReader.java
@@ -282,7 +282,7 @@ public class DFSHoodieDatasetInputReader extends DFSDeltaInputReader {
       return new CloseableMappingIterator<>(reader.getRecordIterator(schema), HoodieRecord::getData);
     } else {
       // If there is no data file, fall back to reading log files
-      HoodieMergedLogRecordScanner scanner = HoodieMergedLogRecordScanner.newBuilder()
+      HoodieMergedLogRecordScanner scanner = (HoodieMergedLogRecordScanner) HoodieMergedLogRecordScanner.newBuilder()
           .withStorage(metaClient.getStorage())
           .withBasePath(metaClient.getBasePath())
           .withLogFilePaths(
